Move useHoverGrow hook to shared project hooks folder
Relocated useHoverGrow from OverviewCard/hooks.js to features/project/hooks/hover-grow.js for better reusability. Updated all imports to reference the new location.
